add apache 2.0 license
I'm not fully sure this is the correct way to do this, but hopefully it's good enough.

Atlassian probably maintains copyright of the original work -- everything in the initial commit in this repo.
The work I've done since then is probably owned by Cru. But I don't really know how to specify this correctly, and I don't think it matters enough to justify more research.